server/requirements.txt
web3
py-solc-x
boto3 
grpcio
grpcio-tools
colorama
threading
